+++ Section26/YelpCamp/routes/comments.js
@@ -1,0 +1,60 @@
+var express = require('express');
+var router = express.Router({
+    mergeParams: true
+});
+var Campground = require('../models/campground');
+var Comment = require('../models/comment');
+
+// ================================
+// COMMENTS NEW 
+// ================================
+router.get("/new", isLoggedIn, (req, res) => {
+    Campground.findById(req.params.id, (err, campground) => {
+        if (err) {
+            console.log(err);
+        } else {
+            res.render("comments/new", {
+                campground: campground
+            });
+        }
+    });
+});
+
+
+// ================================
+// COMMENTS CREATE 
+// ================================
+router.post("/", isLoggedIn, (req, res) => {
+    //lookup campgrounds using ID
+    Campground.findById(req.params.id, (err, campground) => {
+        if (err) {
+            console.log(err);
+            res.redirect("/campgrounds");
+        } else {
+            console.log(req.body.comment);
+            Comment.create(req.body.comment, (err, comment) => {
+                if (err) {
+                    console.log(err);
+                } else {
+                    campground.comments.push(comment);
+                    campground.save();
+                    res.redirect("/campgrounds/" + campground._id);
+                }
+            });
+        }
+    });
+});
+
+
+// ================================
+// MIDDLEWARE
+// ================================
+function isLoggedIn(req, res, next) {
+    if (req.isAuthenticated()) {
+        return next();
+    }
+    res.redirect("/login");
+}
+
+
+module.exports = router;
